Respond right JSON to UI
